Widen __CARDEncode to the 4 bytes its only reader loads (r13/.sdata oracle)
Generalizes the a51d240cf3 width oracle from r2/.sdata2 to r13/.sdata+.sbss.
_SDA_BASE_ = 0x803E31E0 is derived, not assumed: __init_registers in the retail
DOL sets it directly (lis r13,0x803e; ori r13,r13,0x31e0), and it lands exactly
0x8000 above the .sdata start, as _SDA2_BASE_ = 0x803E6500 does for .sdata2.

Decoding all 23449 r13-relative accesses in the retail .text yields ONE symbol
declared smaller than the instruction that reads it -- r13/.sdata is otherwise
clean, which fits: the importer bug behind the 121 .sdata2 fixes was specific to
FP constant atoms, and .sdata/.sbss hold ordinary variables.

  __CARDEncode  size:0x2 data:2byte -> size:0x4 data:4byte  (read by lwz)

Three independent lines agree on 4:
  - the reader: GXTokenInterruptHandler loads it with lwz, not lhz
  - splits.txt already gives CARDBios.c .sbss start:0x803DE0D8 end:0x803DE0DC
  - GXMisc.c already externs it as GXDrawSyncCallback, a 4-byte pointer

That last one made this a live type conflict rather than a cosmetic one: CARDBios.c
DEFINED the slot as a 2-byte u16 while GXMisc.c read it as a 4-byte function
pointer, so our object emitted half the storage its consumer indexes.

The u16 is not a real CARD variable in this build. __CARDEncode is unused by CARD
code -- __CARDGetFontEncode() is defined in __card.h and never called -- and
0x803DE0D8 is not CARD's at all: its .sbss neighbours 0xC0..0xD4 are all GX
globals, and the slot is GXMisc's TokenCB, loaded only by GXTokenInterruptHandler.
CARDBios.c is filling a mis-split slot, so it is widened to u32 rather than renamed.

Restoring the TokenCB name needs the split moved from CARDBios.c to GXMisc.c, and
GXMisc.c's remaining statics (DrawDoneCB/DrawDone/FinishQueue at 0xDC/E0/E4) live
in the auto object, so it cannot own a contiguous .sbss without claiming those too.
Left as a follow-up.

Gated: full ninja EXIT=0, DOL sha1 e750e8e8 unchanged, CARDBios/GXMisc still 100%,
matched_data 1172279 and fuzzy_match_percent 99.83532 unchanged.

+++ include/dolphin/card/__card.h
@@ -64,7 +64,7 @@ s32 __CARDUpdateFatBlock(s32 chan, u16* fat, CARDCallback callback);
 
 // CARDBios
 extern CARDControl __CARDBlock[2];
-extern u16 __CARDEncode;
+extern u32 __CARDEncode;
 
 extern DVDDiskID* __CARDDiskID;
 extern DVDDiskID __CARDDiskNone;
